Given commit is for modifying behaviour of Collector to response of the server.
Additional static method was added to GeneralPerformance class, for
holding sleep time for execution, as it was noticed, that server probably
has some DDoS attack defender which blocks all incomming traffic, if
collecting was too extensive. Waiting is required for resetting new IP
(*as no su rights, later to be changed for changing tor service to restart
every 10 minutes or so).